requirements.txt
flask>=3.0
gunicorn>=22.0
uvicorn>=0.30
azure-identity>=1.17
azure-storage-blob>=12.20
azure-data-tables>=12.5
azure-keyvault-secrets>=4.8
msal>=1.28
requests>=2.31
httpx>=0.27
